refactor: moved Tags.svelte to the same folder of ValueInput.svelte
